requirements.txt
# Django
Django==1.11.3
django-jinja==2.3.1
django-braces==1.11.0

# Database
psycopg2==2.7.1
dj-database-url==0.4.2

# Forms
django-crispy-forms==1.6.1
django-markdown-app==0.9.2
django-taggit==0.22.1
taggit-selectize==2.3.0

# Auth
django-allauth==0.32.0

# Testing
factory-boy==2.9.2
selenium==3.5.0
